fix to code style and organization
